+++ components/TextControls/src/BaseTextAreas/private/MDCBaseTextAreaLayout.m
@@ -15,7 +15,9 @@
 #import "MDCBaseTextAreaLayout.h"
 
 #import "MDCTextControlAssistiveLabelViewLayout.h"
+#import "MDCTextControlHorizontalPositioning.h"
 #import "MDCTextControlLabelPosition.h"
+#import "MDCTextControlSideViewSupport.h"
 #import "MDCTextControlVerticalPositioningReference.h"
 
 #import <MDFInternationalization/MDFInternationalization.h>
@@ -38,6 +40,9 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
 @property(nonatomic, strong, nonnull) NSArray<NSNumber *> *verticalGradientLocations;
 @property(nonatomic, strong, nonnull) NSArray<NSNumber *> *horizontalGradientLocations;
 
+@property(nonatomic, assign) BOOL placeholderLabelHidden;
+@property(nonatomic, assign) CGRect placeholderLabelFrame;
+
 @end
 
 @implementation MDCBaseTextAreaLayout
@@ -53,6 +58,11 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
                                label:(nonnull UILabel *)label
                        labelPosition:(MDCTextControlLabelPosition)labelPosition
                        labelBehavior:(MDCTextControlLabelBehavior)labelBehavior
+                    placeholderLabel:(nonnull UILabel *)placeholderLabel
+                         leadingView:(nullable UIView *)leadingView
+                     leadingViewMode:(UITextFieldViewMode)leadingViewMode
+                        trailingView:(nullable UIView *)trailingView
+                    trailingViewMode:(UITextFieldViewMode)trailingViewMode
                leadingAssistiveLabel:(UILabel *)leadingAssistiveLabel
               trailingAssistiveLabel:(UILabel *)trailingAssistiveLabel
           assistiveLabelDrawPriority:
@@ -71,6 +81,11 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
                                    label:label
                            labelPosition:labelPosition
                            labelBehavior:labelBehavior
+                        placeholderLabel:placeholderLabel
+                             leadingView:leadingView
+                         leadingViewMode:leadingViewMode
+                            trailingView:trailingView
+                        trailingViewMode:trailingViewMode
                    leadingAssistiveLabel:leadingAssistiveLabel
                   trailingAssistiveLabel:trailingAssistiveLabel
               assistiveLabelDrawPriority:assistiveLabelDrawPriority
@@ -92,6 +107,11 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
                                label:(UILabel *)label
                        labelPosition:(MDCTextControlLabelPosition)labelPosition
                        labelBehavior:(MDCTextControlLabelBehavior)labelBehavior
+                    placeholderLabel:(UILabel *)placeholderLabel
+                         leadingView:(UIView *)leadingView
+                     leadingViewMode:(UITextFieldViewMode)leadingViewMode
+                        trailingView:(UIView *)trailingView
+                    trailingViewMode:(UITextFieldViewMode)trailingViewMode
                leadingAssistiveLabel:(UILabel *)leadingAssistiveLabel
               trailingAssistiveLabel:(UILabel *)trailingAssistiveLabel
           assistiveLabelDrawPriority:
@@ -99,19 +119,47 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
     customAssistiveLabelDrawPriority:(CGFloat)customAssistiveLabelDrawPriority
                                isRTL:(BOOL)isRTL
                            isEditing:(BOOL)isEditing {
+  UIView *leftView = isRTL ? trailingView : leadingView;
+  UIView *rightView = isRTL ? leadingView : trailingView;
+  UITextFieldViewMode leftViewMode = isRTL ? trailingViewMode : leadingViewMode;
+  UITextFieldViewMode rightViewMode = isRTL ? leadingViewMode : trailingViewMode;
+
+  BOOL displaysLeftView =
+      MDCTextControlShouldDisplaySideViewWithSideView(leftView, leftViewMode, isEditing);
+  BOOL displaysRightView =
+      MDCTextControlShouldDisplaySideViewWithSideView(rightView, rightViewMode, isEditing);
+
   CGFloat leadingEdgePadding = horizontalPositioningReference.leadingEdgePadding;
   CGFloat trailingEdgePadding = horizontalPositioningReference.trailingEdgePadding;
   CGFloat leftEdgePadding = isRTL ? trailingEdgePadding : leadingEdgePadding;
   CGFloat rightEdgePadding = isRTL ? leadingEdgePadding : trailingEdgePadding;
+  CGFloat horizontalInterItemPadding = horizontalPositioningReference.horizontalInterItemSpacing;
 
-  CGFloat globalTextMinX = leftEdgePadding;
-  CGFloat globalTextMaxX = size.width - rightEdgePadding;
+  CGFloat leftViewWidth = CGRectGetWidth(leftView.frame);
+  CGFloat leftViewMinX = 0;
+  CGFloat leftViewMaxX = 0;
+  if (displaysLeftView) {
+    leftViewMinX = leftEdgePadding;
+    leftViewMaxX = leftViewMinX + leftViewWidth;
+  }
+
+  CGFloat textAreaWidth = size.width;
+  CGFloat rightViewMinX = 0;
+  if (displaysRightView) {
+    CGFloat rightViewMaxX = textAreaWidth - rightEdgePadding;
+    rightViewMinX = rightViewMaxX - CGRectGetWidth(rightView.frame);
+  }
+
+  CGFloat textRectMinX =
+      displaysLeftView ? leftViewMaxX + horizontalInterItemPadding : leftEdgePadding;
+  CGFloat textRectMaxX = displaysRightView ? rightViewMinX - horizontalInterItemPadding
+                                           : textAreaWidth - rightEdgePadding;
 
   CGRect floatingLabelFrame =
       [self floatingLabelFrameWithText:label.text
                                         floatingFont:floatingFont
-                                      globalTextMinX:globalTextMinX
-                                      globalTextMaxX:globalTextMaxX
+                                        textRectMinX:textRectMinX
+                                        textRectMaxX:textRectMaxX
           paddingBetweenContainerTopAndFloatingLabel:verticalPositioningReference
                                                          .paddingBetweenContainerTopAndFloatingLabel
                                                isRTL:isRTL];
@@ -120,8 +168,8 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
   CGRect normalLabelFrame =
       [self normalLabelFrameWithLabelText:label.text
                                               font:font
-                                    globalTextMinX:globalTextMinX
-                                    globalTextMaxX:globalTextMaxX
+                                      textRectMinX:textRectMinX
+                                      textRectMaxX:textRectMaxX
           paddingBetweenContainerTopAndNormalLabel:verticalPositioningReference
                                                        .paddingBetweenContainerTopAndNormalLabel
                                              isRTL:isRTL];
@@ -130,8 +178,6 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
 
   CGFloat textViewMinY = 0;
   CGFloat bottomPadding = 0;
-  CGFloat textViewHeight = 0;
-  CGRect textViewFrame = CGRectZero;
   CGFloat containerHeight = 0;
   BOOL shouldLayoutForFloatingLabel = MDCTextControlShouldLayoutForFloatingLabelWithLabelPosition(
       labelPosition, labelBehavior, label.text);
@@ -145,18 +191,46 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
     }
     bottomPadding = verticalPositioningReference.paddingBetweenEditingTextAndContainerBottom;
     containerHeight = verticalPositioningReference.containerHeightWithFloatingLabel;
-    textViewHeight = containerHeight - bottomPadding - textViewMinY;
-    textViewFrame =
-        CGRectMake(globalTextMinX, textViewMinY, globalTextMaxX - globalTextMinX, textViewHeight);
   } else {
     textViewMinY = verticalPositioningReference.paddingAroundTextWhenNoFloatingLabel;
     bottomPadding = verticalPositioningReference.paddingAroundTextWhenNoFloatingLabel;
     containerHeight = verticalPositioningReference.containerHeightWithoutFloatingLabel;
-    textViewHeight = containerHeight - bottomPadding - textViewMinY;
-    textViewFrame =
-        CGRectMake(globalTextMinX, textViewMinY, globalTextMaxX - globalTextMinX, textViewHeight);
     normalLabelFrame.origin.y = textViewMinY;
   }
+  CGFloat textViewWidth = textRectMaxX - textRectMinX;
+  CGFloat textViewHeight = containerHeight - bottomPadding - textViewMinY;
+  CGRect textViewFrame = CGRectMake(textRectMinX, textViewMinY, textViewWidth, textViewHeight);
+
+  self.placeholderLabelHidden = placeholderLabel.attributedText == nil;
+  if (self.placeholderLabelHidden) {
+    self.placeholderLabelFrame = CGRectZero;
+  } else {
+    CGSize sizeThatFits = [placeholderLabel sizeThatFits:textViewFrame.size];
+    CGFloat placeholderLabelMinX = 0;
+    if (isRTL) {
+      placeholderLabelMinX = textViewWidth - sizeThatFits.width;
+    }
+    self.placeholderLabelFrame =
+        CGRectMake(placeholderLabelMinX, 0, sizeThatFits.width, sizeThatFits.height);
+  }
+
+  CGRect leftViewFrame = CGRectZero;
+  if (displaysLeftView) {
+    leftViewFrame = [self sideViewFrameWithSideView:leftView
+                                               minX:leftViewMinX
+                                    containerHeight:containerHeight];
+  }
+  CGRect rightViewFrame = CGRectZero;
+  if (displaysRightView) {
+    rightViewFrame = [self sideViewFrameWithSideView:rightView
+                                                minX:rightViewMinX
+                                     containerHeight:containerHeight];
+  }
+
+  self.leadingViewFrame = isRTL ? rightViewFrame : leftViewFrame;
+  self.trailingViewFrame = isRTL ? leftViewFrame : rightViewFrame;
+  self.displaysLeadingView = isRTL ? displaysRightView : displaysLeftView;
+  self.displaysTrailingView = isRTL ? displaysLeftView : displaysRightView;
 
   self.assistiveLabelViewLayout = [[MDCTextControlAssistiveLabelViewLayout alloc]
                          initWithWidth:size.width
@@ -222,17 +296,25 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
   return maxY;
 }
 
+- (CGRect)sideViewFrameWithSideView:(UIView *)sideView
+                               minX:(CGFloat)minX
+                    containerHeight:(CGFloat)containerHeight {
+  CGFloat sideViewHeight = CGRectGetHeight(sideView.bounds);
+  CGFloat minY = (0.5f * containerHeight) - (0.5f * sideViewHeight);
+  return CGRectMake(minX, minY, CGRectGetWidth(sideView.bounds), sideViewHeight);
+}
+
 - (CGRect)normalLabelFrameWithLabelText:(NSString *)labelText
                                         font:(UIFont *)font
-                              globalTextMinX:(CGFloat)globalTextMinX
-                              globalTextMaxX:(CGFloat)globalTextMaxX
+                                textRectMinX:(CGFloat)textRectMinX
+                                textRectMaxX:(CGFloat)textRectMaxX
     paddingBetweenContainerTopAndNormalLabel:(CGFloat)paddingBetweenContainerTopAndNormalLabel
                                        isRTL:(BOOL)isRTL {
-  CGFloat maxTextWidth = globalTextMaxX - globalTextMinX;
+  CGFloat maxTextWidth = textRectMaxX - textRectMinX;
   CGSize normalLabelSize = [self textSizeWithText:labelText font:font maxWidth:maxTextWidth];
-  CGFloat normalLabelMinX = globalTextMinX;
+  CGFloat normalLabelMinX = textRectMinX;
   if (isRTL) {
-    normalLabelMinX = globalTextMaxX - normalLabelSize.width;
+    normalLabelMinX = textRectMaxX - normalLabelSize.width;
   }
   CGFloat normalLabelMinY = paddingBetweenContainerTopAndNormalLabel;
   return CGRectMake(normalLabelMinX, normalLabelMinY, normalLabelSize.width,
@@ -241,16 +323,16 @@ static const CGFloat kGradientBlurLength = (CGFloat)4.0;
 
 - (CGRect)floatingLabelFrameWithText:(NSString *)text
                                   floatingFont:(UIFont *)floatingFont
-                                globalTextMinX:(CGFloat)globalTextMinX
-                                globalTextMaxX:(CGFloat)globalTextMaxX
+                                  textRectMinX:(CGFloat)textRectMinX
+                                  textRectMaxX:(CGFloat)textRectMaxX
     paddingBetweenContainerTopAndFloatingLabel:(CGFloat)paddingBetweenContainerTopAndFloatingLabel
                                          isRTL:(BOOL)isRTL {
-  CGFloat maxTextWidth = globalTextMaxX - globalTextMinX;
+  CGFloat maxTextWidth = textRectMaxX - textRectMinX;
   CGSize floatingLabelSize = [self textSizeWithText:text font:floatingFont maxWidth:maxTextWidth];
   CGFloat textMinY = paddingBetweenContainerTopAndFloatingLabel;
-  CGFloat textMinX = globalTextMinX;
+  CGFloat textMinX = textRectMinX;
   if (isRTL) {
-    textMinX = globalTextMaxX - floatingLabelSize.width;
+    textMinX = textRectMaxX - floatingLabelSize.width;
   }
   return CGRectMake(textMinX, textMinY, floatingLabelSize.width, floatingLabelSize.height);
 }