+++ components/TextControls/src/BaseTextFields/MDCBaseTextField.m
@@ -292,6 +292,10 @@ static char *const kKVOContextMDCBaseTextField = "kKVOContextMDCBaseTextField";
     horizontalPositioningReference.trailingEdgePadding =
         (CGFloat)[self.trailingEdgePaddingOverride doubleValue];
   }
+  if (self.horizontalInterItemSpacingOverride) {
+    horizontalPositioningReference.horizontalInterItemSpacing =
+        (CGFloat)[self.horizontalInterItemSpacingOverride doubleValue];
+  }
   return horizontalPositioningReference;
 }