+++ components/TextControls/tests/snapshot/MDCBaseTextAreaSnapshotTests.m
@@ -107,4 +107,90 @@
   [self validateTextArea:textArea];
 }
 
+- (void)testTextAreaWithArabigLabelAndPlaceholderWhileEditingRTL {
+  // Given
+  MDCBaseTextArea *textArea = self.textArea;
+
+  // When
+  textArea.placeholder = @"لكن لا بد";
+  textArea.label.text = @"لكن لا بد";
+  textArea.semanticContentAttribute = UISemanticContentAttributeForceRightToLeft;
+  [textArea.textView becomeFirstResponder];
+
+  // Then
+  [self validateTextArea:textArea];
+}
+
+- (void)testTextAreaWhileEditingWithLabelAndTrailingViewUnlessEditingAndLeadingViewWhileEditing {
+  // Given
+  MDCBaseTextArea *textArea = self.textArea;
+
+  // When
+  textArea.label.text = @"This is a floating label";
+  textArea.leadingView = [self createSideView];
+  textArea.leadingViewMode = UITextFieldViewModeWhileEditing;
+  textArea.trailingView = [self createSideView];
+  textArea.trailingViewMode = UITextFieldViewModeUnlessEditing;
+  textArea.minimumNumberOfVisibleRows = 3;
+  [textArea.textView becomeFirstResponder];
+
+  // Then
+  [self validateTextArea:textArea];
+}
+
+- (void)testTextAreaWhileEditingWithLabelAndLeadingViewUnlessEditingAndTrailingViewWhileEditing {
+  // Given
+  MDCBaseTextArea *textArea = self.textArea;
+
+  // When
+  textArea.label.text = @"This is a floating label";
+  textArea.leadingView = [self createSideView];
+  textArea.leadingViewMode = UITextFieldViewModeUnlessEditing;
+  textArea.trailingView = [self createSideView];
+  textArea.trailingViewMode = UITextFieldViewModeWhileEditing;
+  textArea.minimumNumberOfVisibleRows = 3;
+  [textArea.textView becomeFirstResponder];
+
+  // Then
+  [self validateTextArea:textArea];
+}
+
+- (void)testTextAreaWithPreferredContainerHeightOf300 {
+  // Given
+  MDCBaseTextArea *textArea = self.textArea;
+
+  // When
+  textArea.label.text = @"This is a label";
+  textArea.preferredContainerHeight = 300.0f;
+
+  // Then
+  [self validateTextArea:textArea];
+}
+
+- (void)testTextAreaWithLeadingViewTrailingViewAndCustomPaddings {
+  // Given
+  MDCBaseTextArea *textArea = self.textArea;
+
+  // When
+  textArea.label.text = @"Label text";
+  textArea.leadingView = [self createSideView];
+  textArea.leadingViewMode = UITextFieldViewModeAlways;
+  textArea.leadingView = [self createSideView];
+  textArea.leadingViewMode = UITextFieldViewModeAlways;
+  textArea.leadingEdgePaddingOverride = @(30.0f);
+  textArea.trailingEdgePaddingOverride = @(30.0f);
+  textArea.horizontalInterItemSpacingOverride = @(30.0f);
+
+  // Then
+  [self validateTextArea:textArea];
+}
+
+#pragma mark Helpers
+
+- (UIView *)createSideView {
+  UIView *sideView = [[UIView alloc] initWithFrame:CGRectMake(0, 0, 20, 20)];
+  sideView.backgroundColor = [UIColor blueColor];
+  return sideView;
+}
+
 @end